+++ asd_common/include/response.h
@@ -21,8 +21,8 @@ struct Response
 {
     char    header[RESPONSE_HEADER_SIZE]; // Header identifier string
     char    messages[RESPONSE_MSG_SIZE];  // Error messages for the user
-    int16_t retval;                       // 0: OK, 1: NOT OK
-    int16_t out0;                         // Command output, if any
+    int8_t retval;                        // 0: OK, -1: NOT OK
+    int8_t out0;                          // Command output, -1 if none
     //    Request req;                    // Request being answered
 } __attribute__((__packed__));
